+++ sdk/communication/azure-communication-callautomation/src/main/java/com/azure/communication/callautomation/models/StartRecordingOptions.java
@@ -29,6 +29,8 @@ public class StartRecordingOptions {
 
     private List<CommunicationIdentifier> audioChannelParticipantOrdering;
 
+    private ExternalStorage externalStorage;
+
     /**
      * Constructor
      *
@@ -153,6 +155,26 @@ public class StartRecordingOptions {
      */
     public StartRecordingOptions setAudioChannelParticipantOrdering(List<CommunicationIdentifier> audioChannelParticipantOrdering) {
         this.audioChannelParticipantOrdering = audioChannelParticipantOrdering;
+        return this;
+    }
+
+    /**
+     * Get the externalStorage property: Used to specify external storage for call recording
+     *
+     * @return the externalStorage value.
+     */
+    public ExternalStorage getExternalStorage() {
+        return this.externalStorage;
+    }
+
+    /**
+     * Set the externalStorage property: Used to specify external storage for call recording
+     *
+     * @param externalStorage the external storage for call recording
+     * @return the StartRecordingOptions object itself.
+     */
+    public StartRecordingOptions setExternalStorage(ExternalStorage externalStorage) {
+        this.externalStorage = externalStorage;
         return this;
     }
 }