+++ sdk/spring/spring-cloud-azure-appconfiguration-config/src/test/java/com/azure/spring/cloud/config/implementation/AppConfigurationFeatureManagementPropertySourceTest.java
@@ -35,11 +35,13 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 
+import com.azure.core.util.Configuration;
 import com.azure.data.appconfiguration.models.ConfigurationSetting;
 import com.azure.data.appconfiguration.models.FeatureFlagConfigurationSetting;
 import com.azure.data.appconfiguration.models.FeatureFlagFilter;
 import com.azure.spring.cloud.config.implementation.feature.management.entity.Feature;
 import com.azure.spring.cloud.config.implementation.feature.management.entity.FeatureSet;
+import com.azure.spring.cloud.config.implementation.http.policy.TracingInfo;
 import com.azure.spring.cloud.config.implementation.properties.AppConfigurationProperties;
 import com.azure.spring.cloud.config.implementation.properties.FeatureFlagStore;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -82,7 +84,7 @@ public class AppConfigurationFeatureManagementPropertySourceTest {
 
     @Mock
     private AppConfigurationReplicaClient clientMock;
-    
+
     private FeatureFlagStore featureFlagStore;
 
     @Mock
@@ -124,6 +126,7 @@ public class AppConfigurationFeatureManagementPropertySourceTest {
         when(featureListMock.iterator()).thenReturn(FEATURE_ITEMS.iterator());
         when(clientMock.listSettings(Mockito.any()))
             .thenReturn(featureListMock).thenReturn(featureListMock);
+        when(clientMock.getTracingInfo()).thenReturn(new TracingInfo(false, false, 0, Configuration.getGlobalConfiguration()));
         featureFlagStore.setEnabled(true);
 
         propertySource.initProperties();
@@ -149,6 +152,7 @@ public class AppConfigurationFeatureManagementPropertySourceTest {
     public void testFeatureFlagThrowError() {
         when(featureListMock.iterator()).thenReturn(FEATURE_ITEMS.iterator());
         when(clientMock.listSettings(Mockito.any())).thenReturn(featureListMock);
+        when(clientMock.getTracingInfo()).thenReturn(new TracingInfo(false, false, 0, Configuration.getGlobalConfiguration()));
         try {
             propertySource.initProperties();
         } catch (Exception e) {
@@ -178,6 +182,7 @@ public class AppConfigurationFeatureManagementPropertySourceTest {
         when(featureListMock.iterator()).thenReturn(FEATURE_ITEMS_TARGETING.iterator());
         when(clientMock.listSettings(Mockito.any()))
             .thenReturn(featureListMock).thenReturn(featureListMock);
+        when(clientMock.getTracingInfo()).thenReturn(new TracingInfo(false, false, 0, Configuration.getGlobalConfiguration()));
         featureFlagStore.setEnabled(true);
 
         propertySource.initProperties();