+++ sdk/storage/azure-storage-file-share/src/main/java/com/azure/storage/file/share/ShareFileClient.java
@@ -14,7 +14,7 @@ import com.azure.core.util.Context;
 import com.azure.core.util.FluxUtil;
 import com.azure.core.util.polling.SyncPoller;
 import com.azure.storage.common.ParallelTransferOptions;
-import com.azure.storage.common.StorageSeekableByteChannel;
+import com.azure.storage.common.implementation.StorageSeekableByteChannel;
 import com.azure.storage.common.StorageSharedKeyCredential;
 import com.azure.storage.common.implementation.StorageImplUtils;
 import com.azure.storage.file.share.models.CloseHandlesInfo;
@@ -175,9 +175,9 @@ public class ShareFileClient {
 
         int chunkSize = options.getChunkSize() != null
             ? options.getChunkSize().intValue() : (int) ShareFileAsyncClient.FILE_MAX_PUT_RANGE_SIZE;
-        return new StorageSeekableByteChannel(chunkSize, null /*readBehavior*/,
+        return new StorageSeekableByteChannel(chunkSize,
             new StorageSeekableByteChannelShareFileWriteBehavior(this, options.getRequestConditions(),
-                options.getFileLastWrittenMode()));
+                options.getFileLastWrittenMode()), 0L);
     }
 
     /**
@@ -191,8 +191,8 @@ public class ShareFileClient {
         int chunkSize = configuredChunkSize != null
             ? configuredChunkSize.intValue()
             : (int) ShareFileAsyncClient.FILE_MAX_PUT_RANGE_SIZE;
-        return new StorageSeekableByteChannel(
-            chunkSize, new StorageSeekableByteChannelShareFileReadBehavior(this, conditions), null /*writeBehavior*/);
+        return new StorageSeekableByteChannel(chunkSize,
+            new StorageSeekableByteChannelShareFileReadBehavior(this, conditions), 0L);
     }
 
     /**