+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -3,12 +3,21 @@
 ## 1.9.0-beta.2 (Unreleased)
 
 ### Features Added
+- Added CAE support to service principal authentication.
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.8.1 (2023-03-06)
+
+#### Dependency Updates
+
+- Upgraded `azure-core` from `1.36.0` to version `1.37.0`.
+- Upgraded `azure-core-http-netty` from `1.13.0` to version `1.13.1`.
+- Upgraded `msal4j` from `1.13.4` to version `1.13.5`.
 
 ## 1.9.0-beta.1 (2023-02-08)