+++ sdk/textanalytics/azure-ai-textanalytics/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release History
 
-## 5.3.0-beta.2 (Unreleased)
+## 5.3.0-beta.3 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 5.3.0-beta.2 (2023-03-07)
 
 ### Features Added
 - Added the following methods for performing abstractive summarization and extractive summarization actions:
@@ -16,8 +26,6 @@
 - Changed `BaseResolution` to an abstract class.
 - Removed class `BooleanResolution` and enum value `BooleanResolution` in the class `ResolutionKind`.
 - Renamed `maxSentenceCount` to `sentenceCount` in abstractive summarization.
-
-### Bugs Fixed
 
 ### Other Changes
 - Integrate synchronous workflow for sync clients so that they do not block on async client APIs.